requirements.txt
Flask==0.11.1
azure-storage-file